fix(get_gtdb.py): remove mash for new versions, fix DBDir iterpolation, remove default GTDB-Tk install
Newer versions of GTDB-Tk (>=2.5) do not use mash, so when configuring for this, comments out the GTDBtk_mash line in config_DBs.txt

[DBDir] is now appropriately used in config, was previously not detected due to a newline character.

No longer install GTDB-Tk by default, allow MG-TK to install based on environment yamls. Added --install-tk flag to install a version of GTDB-Tk which is compataible with the GTDB released being configured.